[REFACTOR] moved onGoogleAuth function to LoginScreen
